Fix warnings (#4239)
* Update raylib_api.* by CI

* Fix typecast warnings

---------

+++ src/rcore.c
@@ -2955,7 +2955,7 @@ float GetGamepadAxisMovement(int gamepad, int axis)
     float value = (axis == GAMEPAD_AXIS_LEFT_TRIGGER || axis == GAMEPAD_AXIS_RIGHT_TRIGGER)? -1.0f : 0.0f;
 
     if ((gamepad < MAX_GAMEPADS) && CORE.Input.Gamepad.ready[gamepad] && (axis < MAX_GAMEPAD_AXIS)) {
-        float movement = value < 0.0f ? CORE.Input.Gamepad.axisState[gamepad][axis] : fabs(CORE.Input.Gamepad.axisState[gamepad][axis]);
+        float movement = value < 0.0f ? CORE.Input.Gamepad.axisState[gamepad][axis] : fabsf(CORE.Input.Gamepad.axisState[gamepad][axis]);
 
         // 0.1f = GAMEPAD_AXIS_MINIMUM_DRIFT/DELTA
         if (movement > value + 0.1f) value = CORE.Input.Gamepad.axisState[gamepad][axis];
